requirements.txt
requests==2.31.0
python-dotenv==1.0.0
PyMuPDF==1.23.8  # For PDF loading
python-docx==1.0.1  # For DOCX loading
sentence-transformers==2.2.2  # For embeddings
faiss-cpu==1.7.4  # For vector storage
rank-bm25==0.2.2  # For sparse retrieval
nltk==3.8.1  # For text preprocessing
numpy==1.24.3  # For numerical operations 